Fix major isr bug
The move to exposure records broke finding all calib products
because I forget to update the dataId from the expRecord with
the detector number. Raw finding still work because the **kwargs
were being passed to the .get('raw') but not for the calib existence
lookups. Now the dataId is updated at the start and used consistently
throughout.