Fix package-wide matching bug.
EFD queries have the actual query in the body of a POST request. If
this isn't matched on in full then you get random data back. It is
actually quite amazing that the tests were passing as well as they
were. The bug-fix now means you actually always get the correct
data back from vcr.